ingestor/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
Faker==24.0.0
fastapi==0.110.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.27.1
